refactor: error cleanup
